config-options are always "unsigned short"
git-svn-id: svn+ssh://svn.lighttpd.net/lighttpd/branches/lighttpd-1.4.x@1932 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/mod_proxy.c
@@ -63,7 +63,7 @@ typedef enum {
 
 typedef struct {
 	array *extensions;
-	int debug;
+	unsigned short debug;
 
 	proxy_balance_t balance;
 } plugin_config;
